+++ src/outcome_spaces/power_spectrum.jl
@@ -2,10 +2,11 @@ export PowerSpectrum
 import FFTW
 
 """
-    PowerSpectrum(δ = 0.0) <: OutcomeSpace
+    PowerSpectrum(δ = 0.0, apply_threshold_to_spectrum = false) <: OutcomeSpace
 
 An [`OutcomeSpace`](@ref) based on the power spectrum of a timeseries (amplitude square of
 its Fourier transform). The optional threshold `δ` sets amplitudes below `δ` to zero.
+The optional `apply_threshold_to_spectrum` applies the threshold to the power spectrum if true.
 
 If used with [`probabilities`](@ref), then the spectrum normalized to sum = 1
 is returned as probabilities.
@@ -22,8 +23,9 @@ The outcome space `Ω` for `PowerSpectrum` is the set of frequencies in Fourier 
 should be multiplied with the sampling rate of the signal, which is assumed to be `1`.
 Input `x` is needed for a well-defined [`outcome_space`](@ref).
 """
-@kwdef struct PowerSpectrum{T<:Real} <: OutcomeSpace
+@kwdef struct PowerSpectrum{T<:Real, X<:Bool} <: OutcomeSpace
     δ::T = 0.0
+    apply_threshold_to_spectrum::X = false
 end
 
 function probabilities_and_outcomes(P::PowerSpectrum, x)
@@ -32,12 +34,15 @@ function probabilities_and_outcomes(P::PowerSpectrum, x)
     end
     f = FFTW.rfft(x)
     amp_squared = abs2.(f)
-    if P.δ > 0
+    if P.δ > 0 && P.apply_threshold_to_spectrum
         amp_squared[amp_squared  .< P.δ] .= 0.0
     end
     probs = Probabilities(amp_squared)
     outs = FFTW.rfftfreq(length(x))
     p = Probabilities(probs, outs)
+    if P.δ > 0 && !P.apply_threshold_to_spectrum
+        p = Probabilities([x >= P.δ ? x : 0.0 for x in p], outs)
+    end
     return p, outcomes(p)
 end