+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog is kept with respect to version 0.11 of Entropies.jl. From version v2.
 
 ## 3.9
 
-- Adding a new feature to `PowerSpectrum` estimator that adds a threshold for "small" amplitudes.
+- Adding a new feature to `PowerSpectrum` estimator that adds a threshold for "small" amplitudes. The threshold is applied to the probabilities by default with an option to apply it to the power spectrum.
 
 ## 3.7.4